misc.js -> helpers.js, created utils/metadata.js
